some non prob debugs

+++ src/main/python/bayou/models/non_prob/surrounding_evidences.py
@@ -138,7 +138,7 @@ class SetsOfSets(SetsOfSomething):
             latent_encoding = tf.where(condition, latent_encoding, zeros)
             latent_encoding = tf.reshape(latent_encoding , [config.batch_size * self.max_nums, self.max_depth, config.latent_size])
 
-            count = tf.math.count_nonzero(tf.reduce_sum(latent_encoding, axis=2), axis=1)
+            count = tf.math.count_nonzero(tf.reduce_sum(latent_encoding, axis=2), axis=1, dtype=tf.float32)
             latent_encoding = tf.reduce_sum(latent_encoding, axis=1)/count
 
             latent_encoding = tf.reshape(latent_encoding, [config.batch_size ,self.max_nums, config.latent_size])
